requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2025.3
